refactor: Car 변경에 의한 CarTest 수정
Car 클래스와 VerificationUtil 삭제에 따른 Test 코드 수정을 진행하였습니다.
